// sanity.types.ts
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type LocalizedString = {
  _type: "localizedString";
  en?: string;
  ro?: string;
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type SITE_SETTINGS_QUERY_RESULT = {
  _id: string;
  _type: "siteSettings";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  siteTitle?: LocalizedString;
  description?: LocalizedString;
  logo?: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  phone?: string;
  email?: string;
  address?: LocalizedString;
  socialLinks?: Array<{
    platform?: "facebook" | "instagram" | "linkedin" | "tiktok" | "youtube";
    url?: string;
    _key: string;
  }>;
} | null;